Add makefile. Add tips from mira.

+++ init.sh
@@ -1,6 +1,10 @@
 path=`pwd`
-for x in `ls -a . | egrep -v "\.$" | egrep -v "init.sh"`
+for x in `ls -a . | egrep -v "\.$" | egrep -v "init.sh" | egrep -v "Makefile"`
 do
-    echo $x
-    ln -s $path/$x ~/$x
+    if [ -f ~/$x ]; then
+        echo "File $x already exists"
+    else
+        echo "Creating symbolic link to $x"
+        ln -s $path/$x ~/$x
+    fi
 done
